Implement new block reward scheme for the UNIGRID project
The new reward structure is:
Block       X -   474 999: 8 units
Block 475 000 -   599 999: 6 units
Block 600 000 -   799 999: 4 units
Block 800 000 - 1 000 000: 2 units
Block         > 1 000 000: 1 unit + deflation calculation.

The deflation is calculated as follows:
For every 50 000 blocks the reward is decreased 1% from its
previous value. Hence, this is a cumulative calculation.

This new reward structure will be activated when adoption of the new
daemon and wallet is reached on the network to a satisfactory level.