Add tracking of the blacklisted amount to the "getinfo" command
The daemon will now create the file "blacklistcahce.dat". This cache file
is used to keep track of the blacklisted amount on the chain. Changing
the blacklist spork will mark the blacklist cache as dirty, forcing it to
be recalculated.

The "getinfo" command now also features a "circulatingsupply" command that
simply is the total money supply minus the amount blacklisted on the
blockchain.