Fix mismatch between masternode payments calculation and block height
This is a slight booby trap left behind by James. Basically, the payment
calculation and check of the current block was utilizing the value/amount
of the previous block. Effectively this creates a block with slightly
skewed masternode and development fund payments. At some point when the
block value becomes big enough this would actually stop the chain from
progressing.

Naughty naughty.